Correção na Adção de Contato
Correção na Adção e Remoção de contatos e melhorias no design da página de perfil.

+++ controller/Ligacao.class.php
@@ -3,7 +3,7 @@
 
     class Ligacao {
 
-        public function loadById(int $id_contato,int $id_usuario): LigacaoModel{
+        public function loadById(int $id_usuario,int $id_contato){
             $sql = new Sql();
             $result = $sql->select("SELECT * FROM tb_ligacoes WHERE id_usuario = :ID_USUARIO AND id_contato = :ID_CONTATO", array(
                 ":ID_USUARIO"=>$id_usuario,
@@ -12,8 +12,9 @@
             $ligacao = new LigacaoModel();
             if(count($result)>0){
                 self::setData($ligacao,$result[0]);
+                return $ligacao;
             }
-            return $ligacao;
+            return NULL;
         }
 
         public function loadByUser(int $id,int $limite): array{
@@ -31,18 +32,22 @@
 
         public function add_ligacao(int $id_usuario,int $id_contato){
             $sql = new Sql();
-            $results = $sql->query("INSERT INTO tb_ligacoes(id_usuario, id_contato) VALUES(:ID_USUARIO, :ID_CONTATO)", array(
-                ':ID_USUARIO'=>$id_usuario,
-                ':ID_CONTATO'=>$id_contato
-            ));
+            if($id_usuario != $id_contato){
+                $results = $sql->query("INSERT INTO tb_ligacoes(id_usuario, id_contato) VALUES(:ID_USUARIO, :ID_CONTATO)", array(
+                    ':ID_USUARIO'=>$id_usuario,
+                    ':ID_CONTATO'=>$id_contato
+                ));
+            }
         }
 
         public function rem_ligacao(int $id_usuario,int $id_contato){
             $sql = new Sql();
-            $results = $sql->query("DELETE FROM tb_ligacoes WHERE id_usuario = :ID_USUARIO AND id_contato = :ID_CONTATO", array(
-                ':ID_USUARIO'=>$id_usuario,
-                ':ID_CONTATO'=>$id_contato
-            ));
+            if($id_usuario != $id_contato){
+                $results = $sql->query("DELETE FROM tb_ligacoes WHERE id_usuario = :ID_USUARIO AND id_contato = :ID_CONTATO", array(
+                    ':ID_USUARIO'=>$id_usuario,
+                    ':ID_CONTATO'=>$id_contato
+                ));
+            }
         }
 
         public function setData(LigacaoModel $ligacao,array $data){
